moved test function
