+++ corenlp/src/test/scala/org/clulab/processors/TestBioNLPProcessor2.scala
@@ -2,7 +2,7 @@ package org.clulab.processors
 
 import java.io.{FileReader, BufferedReader}
 
-import org.clulab.odin.TestUtils
+import org.clulab.TestUtils
 import org.clulab.processors.bionlp.BioNLPProcessor
 import org.scalatest._
